refac: 벡엔드 통신 단일화 / 벡엔드 리팩토링 / qml 폴더구조 정리 및 리팩토링
