+++ .gitignore
@@ -4,3 +4,7 @@ __pycache__/
 *$py.class
 venv
 .DS_Store
+
+# 빌드 파일
+dist/
+build/
