--- backend/__init__.py
@@ -1,3 +1,0 @@
-from .initializePortSelect import InitializePortSelect
-from .tooltip import TooltipManager
-from .dock import DockManager, DockableWidget
