+++ backend/serial_manager.py
@@ -1,7 +1,6 @@
-import json
+import time
 import threading
 import serial.tools.list_ports
-import time
 
 from PySide6.QtCore import QObject, Signal, Slot
 
@@ -10,13 +9,9 @@ from .lib.xmlHandler import XmlHandler
 
 
 # QML과 통신을 담당할 백엔드 클래스
-class InitializePortSelect(QObject):
+class SerialManager(QObject):
     connectionResult = Signal(bool, str)  # (성공여부, 메시지)
-
-    connectionSuccessful = Signal(str, int)  # 모니터링 시작을 위한 시그널 (포트, 보드레이트)
-
-    messageMetaDataReady = Signal(str)  # 메시지 메타데이터 전달용 시그널
-    messageUpdated = Signal(list)  # 메시지 업데이트 시그널
+    messageUpdated = Signal(int, dict)  # 메시지 업데이트 시그널
 
     def __init__(self, parent=None):
         super().__init__(parent)
@@ -30,9 +25,14 @@ class InitializePortSelect(QObject):
         self.mav = MiniLink()
         self.data_reading_thread = None
 
+        self.xmlHandler = XmlHandler()
+        self.xmlHandler.loadMessageListFromXML({})
+        self.latest_data = {}
+        self.message_list = []
+
     # QML에서 호출할 수 있는 슬롯 (포트 목록 전달)
     @Slot(result=list)
-    def get_port_list(self):
+    def getPortList(self):
         self.port_list.clear()
         ports = serial.tools.list_ports.comports()
         for port in ports:
@@ -45,7 +45,7 @@ class InitializePortSelect(QObject):
 
     # QML에서 '연결' 버튼을 누르면 호출될 슬롯
     @Slot(str, int)
-    def connect_button_clicked(self, device: str, baudrate: int):
+    def connectButtonClicked(self, device: str, baudrate: int):
         if not device:
             print("오류: 포트가 선택되지 않았습니다.")
             self.connectionResult.emit(False, "포트가 선택되지 않았습니다.")
@@ -62,15 +62,14 @@ class InitializePortSelect(QObject):
             self.port = device
             self.baudrate = baudrate
 
+            self.message_list = self.getMessageList()
+
             # 데이터 읽기 스레드 시작
             self.data_reading_thread = threading.Thread(target=self._get_sensor_data, daemon=True)
             self.data_reading_thread.start()
 
             # 연결 성공 시 QML 변화 trigger
             self.connectionResult.emit(True, f"{device} : 연결 성공")
-
-            # GPS 모니터링 시작 시그널 발생
-            self.connectionSuccessful.emit(device, baudrate)
         except serial.SerialException as e:
             error_msg = f"시리얼 연결 실패: {str(e)}"
             print(error_msg)
@@ -89,28 +88,46 @@ class InitializePortSelect(QObject):
 
         # 연결 확인 코드
         # self.mav.connect() 내부의 serial.Serial()은 시리얼 포트를 여는 코드일 뿐, 실제로 연결됐는지를 보장하지 않음
-        # start_time = time.time()
-        # while True:
-        #     data: list = self.mav.read(enPrint=True, enLog=False)
-        #     if data:
-        #         print("데이터 수신 성공", data)
-        #         break  # 데이터가 성공적으로 읽히면 연결된 것으로 간주
-        #     if time.time() - start_time > 1:  # 1초 동안 데이터가 없으면 연결 실패로 간주
-        #         print("연결 실패: 데이터 수신 대기 시간 초과")
-        #         # raise serial.SerialException("연결 실패: 데이터 수신 대기 시간 초과")
-        #         break
+        # 따라서, 연결 후 3초 동안 데이터 수신이 없으면 연결 실패로 간주
+        self.mav.chooseMessage(26)
+        start_time = time.time()
+        while True:
+            data: list = self.mav.read(enPrint=True, enLog=False)
+            if data:
+                print("연결 성공")
+                break
+            if time.time() - start_time > 3:  # 3초 동안 데이터가 없으면 연결 실패로 간주
+                print("연결 실패")
+                raise serial.SerialException("연결 실패: 데이터 수신 대기 시간 초과")
 
     def _get_sensor_data(self):
         """
         센서 데이터를 지속적으로 읽는 메인 루프
         """
-        self.mav.chooseMessage(26)
 
         try:
+            message_id_list = [msg['id'] for msg in self.message_list]
+            message_frame = {msg['id']: msg['fields'] for msg in self.message_list}
+
+            current_message_idx = 0
+            msg_id = message_id_list[current_message_idx]
+            self.mav.chooseMessage(msg_id)
             while True:
-                data: list = self.mav.read(enPrint=True, enLog=False)
+                data: list = self.mav.read(enPrint=False, enLog=False)
                 if data:
-                    self.messageUpdated.emit(data)
+                    # 데이터 맵핑
+                    msg = {}
+                    for key, value in zip(message_frame[msg_id], data):
+                        msg[key] = value
+                    self.latest_data[msg_id] = msg
+
+                    self.messageUpdated.emit(msg_id, msg)
+
+                    # 다음 메시지 선택
+                    current_message_idx = (current_message_idx + 1) % len(message_id_list)
+                    msg_id = message_id_list[current_message_idx]
+                    self.mav.chooseMessage(msg_id)
+                    # print(f"Latest data: {self.latest_data}")
         except Exception as e:
             print("[Monitor] 연결 끊김 감지!")
             self.port = None
@@ -119,51 +136,24 @@ class InitializePortSelect(QObject):
             return
 
     @Slot(result=dict)
-    def get_current_connection(self):
+    def getCurrentConnection(self):
         return {
             'port': self.port,
             'baudrate': self.baudrate
         }
 
-    @Slot(int)
-    def set_target_message(self, msg_id: int):
-        """
-        QML에서 호출하여 읽을 메시지 ID를 설정합니다.
-        """
-        # 그래프를 그릴 메시지 ID를 설정
-        self.mav.chooseMessage(msg_id)
-
-        # 해당 메시지의 모든 속성을 가져와서 QML에 전달
-        xmlHandler = XmlHandler()
-        xmlHandler.loadMessageListFromXML({})
-        instance = xmlHandler.getMessageInstance(msg_id)
-        fields = [field.attrib for field in instance.findall("field")]
-        for field in fields:
-            field['plot'] = True  # QML에서 플롯팅 여부
-            if 'units' not in field:
-                field['units'] = ''
-
-        meta_data = {
-            'id': msg_id,
-            'name': instance.get("name"),
-            'description': instance.find("description").text,
-            'fields': fields
-        }
-
-        # 시그널로 데이터 전달
-        print(f"메시지 메타데이터 준비 완료: {meta_data}")
-        self.messageMetaDataReady.emit(json.dumps(meta_data))
-
     @Slot(result=list)
-    def get_message_list(self):
+    def getMessageList(self):
         """
         현재 연결된 센서의 메시지 목록을 반환합니다.
         """
         message_list = []
         for key, value in self.mav.getMessageList().items():
+            name = value[0]
+            fields = self.mav.getMessageColumnNames(key)
             message_list.append({
                 'id': key,
-                'name': value[0]
+                'name': name,
+                'fields': fields
             })
-
         return message_list
