+++ backend/tooltip_manager.py
