+++ tests/waypoint-samples/interpolated_sample.txt
