+++ tests/waypoint-samples/waypoint_sample.txt
