+++ windows/manual_gps_window.py
@@ -1,5 +1,5 @@
-from PySide6.QtWidgets import QMainWindow
 from PySide6.QtCore import QUrl
+from PySide6.QtWidgets import QMainWindow
 from PySide6.QtQuickWidgets import QQuickWidget
 
 from backend.utils import resource_path
@@ -14,9 +14,9 @@ class ManualGpsWindow(QMainWindow):
         self.qml_widget = QQuickWidget()
 
         # 소스 로드 전에 컨텍스트 설정
-        self.qml_widget.rootContext().setContextProperty("gpsBackend", gps_backend)
+        self.qml_widget.rootContext().setContextProperty("gpsManager", gps_backend)
 
-        qml_file = resource_path("src/pages/flight/nd/ManualGpsInput.qml")
+        qml_file = resource_path("src/pages/flight/nd/windows/ManualGpsInput.qml")
         self.qml_widget.setSource(QUrl.fromLocalFile(qml_file))
         self.qml_widget.setResizeMode(QQuickWidget.SizeRootObjectToView)
         self.setCentralWidget(self.qml_widget)
